Centralize sentry and stub for sentry features
