+++ backend/src/services/mod.rs
@@ -4,3 +4,10 @@ pub mod genai;
 // pub mod mcp_servers;
 pub mod mcp_manager;
 pub mod mcp_servers_sse;
+
+#[cfg(feature = "sentry")]
+pub mod sentry;
+#[cfg(not(feature = "sentry"))]
+pub mod sentry_stub;
+#[cfg(not(feature = "sentry"))]
+pub use sentry_stub as sentry;
